refactor: aggregate rpc requests
