+++ packages/plugins/ArtBlocks/src/SNSAdaptor/DetailsView.tsx
@@ -1,11 +1,11 @@
 import { FormattedAddress } from '@masknet/shared'
+import { BigNumber } from 'bignumber.js'
 import { makeStyles } from '@masknet/theme'
 import { formatEthereumAddress, formatWeiToEther, chainResolver, explorerResolver } from '@masknet/web3-shared-evm'
 import { useChainContext } from '@masknet/web3-hooks-base'
 import type { NetworkPluginID } from '@masknet/shared-base'
 import { OpenInNew } from '@mui/icons-material'
 import { Typography, Box, Link } from '@mui/material'
-import { BigNumber } from 'bignumber.js'
 import { resolveProjectLinkOnArtBlocks, resolveUserLinkOnArtBlocks } from '../pipes/index.js'
 import type { Project } from '../types.js'
 import { useI18N } from '../locales/index.js'
