+++ packages/plugins/Snapshot/src/SNSAdaptor/Vote.tsx
@@ -1,7 +1,7 @@
 import { Avatar, Box } from '@mui/material'
 import { makeStyles } from '@masknet/theme'
-import type { VoteItem as VoteType } from '../types.js'
 import { useBlockie } from '@masknet/web3-hooks-base'
+import type { VoteItem as VoteType } from '../types.js'
 
 const useStyles = makeStyles()({
     avatar: {
