+++ packages/web3-providers/src/SmartPay/libs/ContractWallet.ts
@@ -1,13 +1,11 @@
 import Web3 from 'web3'
 import type { AbiItem } from 'web3-utils'
 import * as ABICoder from 'web3-eth-abi'
+import { createContract, type ChainId, ZERO_ADDRESS, getSmartPayConstants } from '@masknet/web3-shared-evm'
 import type { WalletProxy } from '@masknet/web3-contracts/types/WalletProxy.js'
 import WalletABI from '@masknet/web3-contracts/abis/Wallet.json'
 import WalletProxyABI from '@masknet/web3-contracts/abis/WalletProxy.json'
 import { WalletProxyByteCode } from '@masknet/web3-contracts/bytes/WalletProxy.mjs'
-import { createContract } from '../helpers/createContract.js'
-import { getSmartPayConstants, ZERO_ADDRESS } from '../constants/index.js'
-import type { ChainId } from '../types/index.js'
 
 export class ContractWallet {
     private web3 = new Web3()
