+++ packages/web3-providers/src/Web3/EVM/interceptors/Popups.ts
@@ -6,7 +6,6 @@ import {
     type TransactionOptions,
     getMaskTokenAddress,
     getSmartPayConstants,
-    DepositPaymaster,
     PayloadEditor,
     EthereumMethodType,
     ProviderType,
@@ -14,6 +13,7 @@ import {
 } from '@masknet/web3-shared-evm'
 import { ExtensionSite, getSiteType, isEnhanceableSiteType } from '@masknet/shared-base'
 import { RequestStateType, isGreaterThan, isZero, toFixed } from '@masknet/web3-shared-base'
+import { DepositPaymaster } from '../../../SmartPay/libs/DepositPaymaster.js'
 import { SharedContextRef } from '../../../PluginContext/index.js'
 import { SmartPayBundlerAPI } from '../../../SmartPay/index.js'
 import { ConnectionReadonlyAPI } from '../apis/ConnectionReadonlyAPI.js'
