+++ packages/web3-providers/src/entry-helpers.ts
@@ -11,5 +11,10 @@ export * from './helpers/getPaymentToken.js'
 export * from './helpers/parseJSON.js'
 export * from './helpers/resolveActivityType.js'
 export * from './helpers/resolveActivityTypeBackgroundColor.js'
+export * from './helpers/createWeb3FromURL.js'
+export * from './helpers/createWeb3FromProvider.js'
+export * from './helpers/createWeb3ProviderFromRequest.js'
+export * from './helpers/createWeb3ProviderFromURL.js'
 
 export * as uniswap from './Trader/helpers/uniswap.js'
+export * as trending from './Trending/helpers.js'
