+++ packages/web3-providers/tests/SmartPay/libs/Create2Factory.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest'
-import { isValidAddress } from '../../src/helpers/address.js'
-import { Create2Factory } from '../../src/libs/Create2Factory.js'
+import { isValidAddress } from '@masknet/web3-shared-evm'
+import { Create2Factory } from '../../../src/SmartPay/libs/Create2Factory.js'
 
 describe('Create2Factory', () => {
     const create2Factory = new Create2Factory('0x0000000000000000000000000000000000000000')
