+++ packages/web3-providers/tests/SmartPay/libs/UserTransaction.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect } from 'vitest'
-import { UserTransaction } from '../../src/libs/UserTransaction.js'
-import { ChainId } from '../../src/types/index.js'
+import { ChainId } from '@masknet/web3-shared-evm'
+import { UserTransaction } from '../../../src/SmartPay/libs/UserTransaction.js'
 
 describe('UserTransaction', async () => {
     const userTransaction = new UserTransaction(ChainId.Matic, '0x0000000000000000000000000000000000000000', {
