+++ packages/web3-shared/evm/src/index.ts
@@ -6,7 +6,7 @@ export * from './helpers/defaults.js'
 export * from './helpers/token.js'
 export * from './helpers/resolver.js'
 export * from './helpers/formatter.js'
-export * from './helpers/provider.js'
+export * from './helpers/rpc.js'
 export * from './helpers/addGasMargin.js'
 export * from './helpers/createContract.js'
 export * from './helpers/decodeOutputString.js'
