+++ packages/web3-shared/evm/src/libs/index.ts
@@ -1,9 +1,6 @@
 export * from './AccountTransaction.js'
 export * from './Composer.js'
 export * from './ContractTransaction.js'
-export * from './ContractWallet.js'
-export * from './Create2Factory.js'
-export * from './DepositPaymaster.js'
 export * from './EIP155Editor.js'
 export * from './EIP3085Editor.js'
 export * from './ErrorEditor.js'
@@ -13,4 +10,3 @@ export * from './PayloadEditor.js'
 export * from './ProviderURL.js'
 export * from './RequestID.js'
 export * from './Signer.js'
-export * from './UserTransaction.js'
